# docs/swagger.yaml
definitions:
  models.Achievement:
    properties:
      achievementType:
        type: string
      attachments:
        items:
          $ref: '#/definitions/models.Attachment'
        type: array
      createdAt:
        type: string
      customFields:
        additionalProperties: true
        type: object
      deletedAt:
        type: string
      description:
        type: string
      details:
        $ref: '#/definitions/models.DynamicDetails'
      id:
        type: string
      is_deleted:
        type: boolean
      points:
        type: number
      studentId:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    required:
    - achievementType
    - studentId
    - title
    type: object
...
  models.AchievementReference:
    properties:
      createdAt:
        type: string
      id:
        type: string
      mongoAchievementId:
        type: string
      rejectionNote:
        type: string
      status:
        $ref: '#/definitions/models.AchievementStatus'
      studentId:
        type: string
      submittedAt:
        type: string
      updatedAt:
        type: string
      verifiedAt:
        type: string
      verifiedBy:
        type: string
    type: object
...
  models.AchievementStatus:
    enum:
    - draft
    - submitted
    - verified
    - rejected
    - deleted
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusSubmitted
    - StatusVerified
    - StatusRejected
    - StatusDeleted
...
  models.Attachment:
    properties:
      fileName:
        type: string
      fileType:
        type: string
      fileUrl:
        type: string
      uploadedAt:
        type: string
    required:
    - fileName
    - fileUrl
    type: object
...
  models.DynamicDetails:
    properties:
      authors:
        items:
          type: string
        type: array
      certificationName:
        type: string
      certificationNumber:
        type: string
      competitionLevel:
        type: string
      competitionName:
        type: string
      eventDate:
        type: string
      issn:
        type: string
      issuedBy:
        type: string
      location:
        type: string
      medalType:
        type: string
      organizationName:
        type: string
      organizer:
        type: string
      period:
        $ref: '#/definitions/models.Period'
      position:
        type: string
      publicationTitle:
        type: string
      publicationType:
        type: string
      publisher:
        type: string
      rank:
        type: integer
      score:
        type: number
      validUntil:
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      data:
        properties:
          refreshToken:
            type: string
          token:
            type: string
          user:
            $ref: '#/definitions/models.UserData'
        type: object
      status:
        type: string
    type: object
  models.Period:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
  models.RefreshTokenResponse:
    properties:
      data:
        properties:
          refreshToken:
            type: string
          token:
            type: string
        type: object
      status:
        type: string
    type: object
...
  models.UserData:
    properties:
      fullName:
        type: string
      id:
        type: string
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API Server untuk manajemen dan pelaporan prestasi mahasiswa.
  title: Sistem Pelaporan Prestasi API
  version: "1.0"
paths:
  /api/v1/achievements:
    get:
      consumes:
      - application/json
      description: Get a list of achievement references for the logged-in student
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AchievementReference'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Achievements
      tags:
      - Achievements
    post:
      consumes:
      - application/json
      description: Create a new achievement record (Mahasiswa only)
      parameters:
      - description: Achievement Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Achievement'
      produces:
      - application/json
      responses:
        "201":
          description: Returns Mongo ID
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create Achievement
      tags:
      - Achievements
  /api/v1/achievements/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete achievement from system (Mahasiswa only)
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Achievement
      tags:
      - Achievements
    get:
      consumes:
      - application/json
      description: Retrieve full details of an achievement from MongoDB
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Achievement'
      security:
      - ApiKeyAuth: []
      summary: Get Achievement Detail
      tags:
      - Achievements
    put:
      consumes:
      - application/json
      description: Update draft achievement details (Mahasiswa only)
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Achievement Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Achievement'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update Achievement
      tags:
      - Achievements
  /api/v1/achievements/{id}/attachments:
    post:
      consumes:
      - application/json
      description: Upload/Add file URL to an achievement
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Attachment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add Attachment
      tags:
      - Achievements
  /api/v1/achievements/{id}/history:
    get:
      consumes:
      - application/json
      description: Retrieve status history and metadata of an achievement
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AchievementReference'
      security:
      - ApiKeyAuth: []
      summary: Get Achievement History
      tags:
      - Achievements
  /api/v1/achievements/{id}/reject:
    post:
      consumes:
      - application/json
      description: Reject achievement with a note (Dosen Wali only)
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection note
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Reject Achievement
      tags:
      - Achievements
  /api/v1/achievements/{id}/submit:
    post:
      consumes:
      - application/json
      description: Change status from Draft to Submitted for verification
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Submit Achievement
      tags:
      - Achievements
  /api/v1/achievements/{id}/verify:
    post:
      consumes:
      - application/json
      description: Approve achievement and assign points (Dosen Wali only)
      parameters:
      - description: Mongo Achievement ID
        in: path
        name: id
        required: true
        type: string
      - description: Points data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Verify Achievement
      tags:
      - Achievements
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access & refresh tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current refresh token
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: User Logout
      tags:
      - Authentication
  /api/v1/auth/profile:
    get:
      consumes:
      - application/json
      description: Retrieve the logged-in user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserData'
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Current Profile
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RefreshTokenResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Refresh Token
      tags:
      - Authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"